+++ .eslintignore
@@ -1,3 +1,3 @@
 /lib
 /docs
-defaultConfig.js
+defaultConfig.stub.js
