+++ __tests__/importsConfig.test.js
@@ -3,5 +3,5 @@ import config from '../defaultConfig.js'
 
 test('it can accept a config file', () => {
   tailwind('./defaultConfig.js')
-  expect(tailwind.defaultConfig()).toEqual(config)
+  expect(require('../defaultConfig.stub.js')).toEqual(config())
 })
