+++ src/cli.js
@@ -52,7 +52,7 @@ program
       process.exit(1)
     }
 
-    const output = fs.readFileSync(path.resolve(__dirname, '../defaultConfig.js'), 'utf8')
+    const output = fs.readFileSync(path.resolve(__dirname, '../defaultConfig.stub.js'), 'utf8')
     fs.outputFileSync(destination, output.replace('// var defaultConfig', 'var defaultConfig'))
     console.log(`Generated Tailwind config: ${destination}`)
     process.exit()
